忽略自定义权限

+++ project/Libraries/DroidPlugin/src/com/morgoo/droidplugin/pm/IPluginManagerImpl.java
@@ -846,9 +846,11 @@ public class IPluginManagerImpl extends IPluginManager.Stub {
                         } catch (NameNotFoundException e) {
                         }
                         if (!mHostRequestedPermission.contains(requestedPermission) && b) {
-                            Log.e(TAG, "No Permission %s", requestedPermission);
-                            new File(apkfile).delete();
-                            return PluginManager.INSTALL_FAILED_NO_REQUESTEDPERMISSION;
+                            if(!PluginManager.IGNORE_CUSTOM_PERMISSION) {
+                                Log.e(TAG, "No Permission %s", requestedPermission);
+                                new File(apkfile).delete();
+                                return PluginManager.INSTALL_FAILED_NO_REQUESTEDPERMISSION;
+                            }
                         }
                     }
                 }
@@ -887,8 +889,10 @@ public class IPluginManagerImpl extends IPluginManager.Stub {
                             }
                             if (!mHostRequestedPermission.contains(requestedPermission) && b) {
                                 Log.e(TAG, "No Permission %s", requestedPermission);
-                                new File(apkfile).delete();
-                                return PluginManager.INSTALL_FAILED_NO_REQUESTEDPERMISSION;
+                                if(!PluginManager.IGNORE_CUSTOM_PERMISSION) {
+                                    new File(apkfile).delete();
+                                    return PluginManager.INSTALL_FAILED_NO_REQUESTEDPERMISSION;
+                                }
                             }
                         }
                     }
@@ -924,7 +928,7 @@ public class IPluginManagerImpl extends IPluginManager.Stub {
         String packageName = parser.getPackageName();
         String optimizedDirectory = PluginDirHelper.getPluginDalvikCacheDir(hostContext, packageName);
         String libraryPath = PluginDirHelper.getPluginNativeLibraryDir(hostContext, packageName);
-        ClassLoader classloader = new PluginClassLoader(apkfile, optimizedDirectory, libraryPath,hostContext.getClassLoader().getParent());
+        ClassLoader classloader = new PluginClassLoader(apkfile, optimizedDirectory, libraryPath, hostContext.getClassLoader().getParent());
 //        DexFile dexFile = DexFile.loadDex(apkfile, PluginDirHelper.getPluginDalvikCacheFile(mContext, parser.getPackageName()), 0);
 //        Log.e(TAG, "dexFile=%s,1=%s,2=%s", dexFile, DexFile.isDexOptNeeded(apkfile), DexFile.isDexOptNeeded(PluginDirHelper.getPluginDalvikCacheFile(mContext, parser.getPackageName())));
     }
